requirements.txt
langchain==0.3.14
langchain-community
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langgraph==0.2.62
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.2.11
loguru==0.7.3
langchain-groq==0.3.1 
python-dotenv==1.1.0
streamlit==1.33.0
pdfplumber==0.10.3
python-docx==1.1.0
faiss-cpu==1.7.4
unstructured==0.12.5
sentence-transformers==2.6.2
pdfplumber==0.11.5
python-docx==1.1.2
pypdf==5.4.0 
faiss-cpu==1.10.0 